Move matplotlib functionality into backend
